requirements.txt
torch==2.0.0
torchvision
opencv-python
pillow
streamlit
pandas
sckitlearn
